+++ src/core/models/Correspondent.py
@@ -29,6 +29,7 @@ import httpx
 from django.conf import settings
 from django.db import models
 from django.utils.translation import gettext_lazy as _
+from django_prometheus.models import ExportModelOperationsMixin
 from rest_framework import status
 from vobject import vCard
 
@@ -51,7 +52,12 @@ logger = logging.getLogger(__name__)
 
 
 class Correspondent(
-    DownloadMixin, URLMixin, FavoriteModelMixin, TimestampModelMixin, models.Model
+    ExportModelOperationsMixin("correspondent"),
+    DownloadMixin,
+    URLMixin,
+    FavoriteModelMixin,
+    TimestampModelMixin,
+    models.Model,
 ):
     """Database model for the correspondent data found in a mail."""