+++ src/core/models/Daemon.py
@@ -31,6 +31,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 from django_celery_beat.models import IntervalSchedule, PeriodicTask
+from django_prometheus.models import ExportModelOperationsMixin
 
 from core.constants import EmailFetchingCriterionChoices
 from core.mixins import HealthModelMixin, TimestampModelMixin, URLMixin
@@ -45,7 +46,12 @@ logger = logging.getLogger(__name__)
 
 
 class Daemon(
-    DirtyFieldsMixin, URLMixin, HealthModelMixin, TimestampModelMixin, models.Model
+    ExportModelOperationsMixin("routine"),
+    DirtyFieldsMixin,
+    URLMixin,
+    HealthModelMixin,
+    TimestampModelMixin,
+    models.Model,
 ):
     """Database model for the daemon fetching a mailbox.